Rename Alert module to Notice
